requirements.txt
h5py
matplotlib
numpy
Pillow
tensorflow==2.6.0
tqdm
scikit-image
